requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=1.3.0=pypi_0
anyio=3.6.1=pypi_0
aom=3.5.0=h27087fc_0
appdirs=1.4.4=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
ase=3.22.1=pypi_0
attrs=22.1.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
black=22.10.0=pypi_0
blas=1.0=mkl
bleach=5.0.1=pypi_0
blis=0.2.4=pypi_0
boto3=1.26.70=pypi_0
botocore=1.29.70=pypi_0
brotlipy=0.7.0=py37h540881e_1004
bzip2=1.0.8=h7f98852_4
ca-certificates=2023.01.10=h06a4308_0
cachetools=4.2.4=pypi_0
certifi=2022.12.7=py37h06a4308_0
cffi=1.15.1=py37h43b0acd_0
cfgv=3.3.1=pypi_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
chatterbot=1.1.0=pypi_0
click=8.1.3=pypi_0
colorlover=0.3.0=pypi_0
community=1.0.0b1=pypi_0
conda-pack=0.6.0=pyhd3eb1b0_0
cryptography=37.0.1=py37h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cufflinks=0.17.3=pypi_0
cycler=0.11.0=pypi_0
cymem=2.0.7=pypi_0
debugpy=1.6.3=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
distlib=0.3.6=pypi_0
einops=0.6.0=pypi_0
entrypoints=0.4=pypi_0
et-xmlfile=1.1.0=pypi_0
expat=2.4.9=h27087fc_0
faiss-gpu=1.7.2=pypi_0
fastjsonschema=2.16.2=pypi_0
ffmpeg=5.1.2=gpl_he10e716_101
filelock=3.8.0=pypi_0
flask=2.2.2=pypi_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.0=hc2a2eb6_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.37.4=pypi_0
freetype=2.12.1=hca18f0e_0
fuzzywuzzy=0.18.0=pypi_0
gettext=0.21.1=h27087fc_0
gmp=6.2.1=h58526e2_0
gnutls=3.7.8=hf3e180e_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-speech=1.1.0=pypi_0
googledrivedownloader=0.4=pypi_0
grpcio=1.49.1=pypi_0
h5py=3.7.0=pypi_0
icu=70.1=h27087fc_0
identify=2.5.6=pypi_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=5.0.0=pypi_0
importlib-resources=5.10.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.16.0=pypi_0
ipython=7.34.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.2=pypi_0
isodate=0.6.1=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.2=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=h166bdaf_2
jsonschema=4.16.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.4.2=pypi_0
jupyter-console=6.4.4=pypi_0
jupyter-core=4.11.1=pypi_0
jupyter-server=1.21.0=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.3=pypi_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h166bdaf_1003
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=4.0.0=h27087fc_0
libblas=3.9.0=12_linux64_mkl
libcblas=3.9.0=12_linux64_mkl
libdeflate=1.14=h166bdaf_0
libdrm=2.4.113=h166bdaf_0
libffi=3.4.2=h7f98852_5
libgcc-ng=12.1.0=h8d9b700_16
libgomp=12.1.0=h8d9b700_16
libiconv=1.17=h166bdaf_0
libidn2=2.3.3=h166bdaf_0
liblapack=3.9.0=12_linux64_mkl
libnsl=2.0.0=h7f98852_0
libpciaccess=0.16=h516909a_0
libpng=1.6.38=h753d276_0
libsqlite=3.39.4=h753d276_0
libstdcxx-ng=12.1.0=ha89aaad_16
libtasn1=4.19.0=h166bdaf_0
libtiff=4.4.0=h55922b4_4
libunistring=0.9.10=h7f98852_0
libuuid=2.32.1=h7f98852_1000
libva=2.16.0=h166bdaf_0
libvpx=1.11.0=h9c3ff4c_3
libwebp-base=1.2.4=h166bdaf_0
libxcb=1.13=h7f98852_1004
libxml2=2.9.14=h22db469_4
libzlib=1.2.13=h166bdaf_4
littleutils=0.2.2=pypi_0
llvmlite=0.39.1=pypi_0
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
mathparse=0.1.2=pypi_0
matplotlib=3.1.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=2.0.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h402132d_0
mkl_fft=1.3.1=py37h3e078e5_1
mkl_random=1.2.2=py37h219a48f_0
more-itertools=9.0.0=pypi_0
murmurhash=1.0.9=pypi_0
mypy-extensions=0.4.3=pypi_0
nbclassic=0.4.5=pypi_0
nbclient=0.7.0=pypi_0
nbconvert=7.2.1=pypi_0
nbformat=5.7.0=pypi_0
ncurses=6.3=h27087fc_1
nest-asyncio=1.5.6=pypi_0
nettle=3.8.1=hc379101_1
networkx=2.6.3=pypi_0
nltk=3.8.1=pypi_0
nodeenv=1.7.0=pypi_0
notebook=6.5.1=pypi_0
notebook-shim=0.2.0=pypi_0
numba=0.56.3=pypi_0
numpy=1.19.4=pypi_0
oauthlib=3.2.1=pypi_0
ogb=1.3.3=pypi_0
opencv-python=4.6.0.66=pypi_0
openh264=2.3.1=h27087fc_1
openjpeg=2.5.0=h7d73246_1
openpyxl=3.1.2=pypi_0
openssl=3.1.0=hd590300_3
outdated=0.2.2=pypi_0
p11-kit=0.24.1=hc5aa10d_0
packaging=21.3=pypi_0
pandas=0.25.1=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathspec=0.10.1=pypi_0
pbr=5.11.1=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.2.0=py37h850a105_2
pint=0.18=pypi_0
pip=23.1.2=pypi_0
pkgutil-resolve-name=1.3.10=pypi_0
plac=0.9.6=pypi_0
platformdirs=2.5.2=pypi_0
plotly=5.10.0=pypi_0
plyfile=0.7.4=pypi_0
pre-commit=2.20.0=pypi_0
preshed=2.0.1=pypi_0
prometheus-client=0.15.0=pypi_0
prompt-toolkit=3.0.31=pypi_0
protobuf=3.19.6=pypi_0
psutil=5.9.3=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pygame=2.3.0=pypi_0
pygments=2.13.0=pypi_0
pyopenssl=22.0.0=pyhd8ed1ab_1
pyparsing=3.0.9=pypi_0
pyrsistent=0.18.1=pypi_0
pyserial=3.5=pypi_0
pysmiles=1.0.2=pypi_0
pysocks=1.7.1=py37h89c1867_5
pytdc=0.3.9=pypi_0
python=3.7.12=hf930737_100_cpython
python-dateutil=2.8.2=pypi_0
python-vlc=3.0.18121=pypi_0
python_abi=3.7=2_cp37m
pytorch=1.12.1=py3.7_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytz=2022.5=pypi_0
pyyaml=5.3.1=pypi_0
pyzmq=24.0.1=pypi_0
qtconsole=5.3.2=pypi_0
qtpy=2.2.1=pypi_0
rdflib=6.2.0=pypi_0
rdkit-pypi=2022.9.4=pypi_0
readline=8.1.2=h0f457ee_0
regex=2022.10.31=pypi_0
requests=2.28.1=pyhd8ed1ab_1
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
s3transfer=0.6.0=pypi_0
scikit-learn=0.21.3=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.12.2=pypi_0
send2trash=1.8.0=pypi_0
setuptools=65.5.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
spacy=2.1.9=pypi_0
speechrecognition=3.9.0=pypi_0
sqlalchemy=1.3.24=pypi_0
sqlite=3.39.4=h4ff8645_0
srsly=1.0.6=pypi_0
stanfordcorenlp=3.9.1.1=pypi_0
svt-av1=1.2.1=h27087fc_0
tdc=0.1=pypi_0
tenacity=8.1.0=pypi_0
tensorboard=2.0.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
terminado=0.16.0=pypi_0
thinc=7.0.8=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h27826a3_0
toml=0.10.2=pypi_0
tomli=2.0.1=pypi_0
torch-cluster=1.6.0=pypi_0
torch-geometric=2.1.0.post1=pypi_0
torch-scatter=2.0.9=pypi_0
torch-sparse=0.6.14=pypi_0
torch-spline-conv=1.2.1=pypi_0
torchaudio=0.12.1=py37_cu113
torchvision=0.13.1=py37_cu113
tornado=6.2=pypi_0
tqdm=4.64.1=pypi_0
traitlets=5.5.0=pypi_0
typed-ast=1.5.4=pypi_0
typing_extensions=4.4.0=pyha770c72_0
urllib3=1.26.11=pyhd8ed1ab_0
utils=1.0.1=pypi_0
virtualenv=20.16.5=pypi_0
wasabi=0.10.1=pypi_0
wcwidth=0.2.5=pypi_0
web-cache=1.1.0=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.4.1=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=4.0.3=pypi_0
wikipedia=1.4.0=pypi_0
x264=1!164.3095=h166bdaf_2
x265=3.5=h924138e_3
xlrd=1.2.0=pypi_0
xorg-fixesproto=5.0=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.6=h166bdaf_0
zipp=3.9.0=pypi_0
zstd=1.5.2=h6239696_4
